Can now compile on windows with MSVC. Changed some msvc flags, If no CMAKE_BUILD_TYPE is specified, now Debug will be default

+++ reviv/src/core/time.h
@@ -21,7 +21,7 @@ public:
     float getTime(); // returns time since the start of rendering
     bool isOneSecond();
 
-    float getLoopCounter() { return get()->gameLoopCounter; }
+    unsigned long long getLoopCounter() { return get()->gameLoopCounter; }
 
     float timeRatio = 1.f; // used for speeding up or slowing down game time
